playlist_classifier/requirements.txt
Flask
spotipy
scikit-learn
pandas
numpy
joblib
python-dotenv